+++ src/organisms/SubmissionsPageTable/index.js
@@ -1,0 +1,3 @@
+import SubmissionsPageTableContainer from "./SubmissionsPageTableContainer";
+
+export default SubmissionsPageTableContainer;
